Move marker handling out of main.js
This is a necessary refactor to be able to properly remeasure block
decorations in Atom 1.19.

As a added bonus we can now slim down main.js a lot since all storage is
now handled inside store. This will allow us to make our code better
testable and easier to read.