Provide ordering to configuration settings (#1530)
Hydrogen has grown to the point where there are 14 configuration settings. Currently, no ordering is provided, so I think they're ordered alphabetically in terms of the config key. I think this is more confusing than it has to be for users.

This PR attempts to group settings logically and then order them with the `order` value in the dictionary. The idea is "general settings" then "output settings" then "kernel settings", and within each ordered by most used to least used. 

The order is:
- `autocomplete`
- `showAutocompleteFirst`
- `statusBarDisable`
- `debug`
- `autoScroll`
- `outputAreaDefault`
- `outputAreaDock`
- `outputAreaFontSize`
- `globalMode`
- `kernelNotifications`
- `startDir`
- `languageMappings`
- `startupCode`
- `gateways`